polyfilled legacy described endpoints
