+++ nym-api/src/nym_contract_cache/handlers.rs
@@ -6,68 +6,14 @@ use crate::node_status_api::helpers::{
     _get_rewarded_set_legacy_mixnodes_detailed,
 };
 use crate::support::http::state::AppState;
+use crate::support::legacy_helpers::{to_legacy_gateway, to_legacy_mixnode};
 use axum::extract::State;
 use axum::{Json, Router};
-use nym_api_requests::legacy::{LegacyMixNodeBondWithLayer, LegacyMixNodeDetailsWithLayer};
-use nym_api_requests::models::{MixNodeBondAnnotated, NymNodeData};
-use nym_config::defaults::DEFAULT_NYM_NODE_HTTP_PORT;
-use nym_mixnet_contract_common::mixnode::LegacyPendingMixNodeChanges;
+use nym_api_requests::legacy::LegacyMixNodeDetailsWithLayer;
+use nym_api_requests::models::MixNodeBondAnnotated;
 use nym_mixnet_contract_common::reward_params::Performance;
-use nym_mixnet_contract_common::{
-    reward_params::RewardingParams, Gateway, GatewayBond, Interval, LegacyMixLayer, MixNode,
-    MixNodeBond, NodeId, NymNodeDetails,
-};
-use rand::prelude::SliceRandom;
-use rand::rngs::OsRng;
+use nym_mixnet_contract_common::{reward_params::RewardingParams, GatewayBond, Interval, NodeId};
 use std::collections::HashSet;
-
-fn to_legacy_mixnode(
-    nym_node: &NymNodeDetails,
-    description: &NymNodeData,
-) -> LegacyMixNodeDetailsWithLayer {
-    let layer_choices = [
-        LegacyMixLayer::One,
-        LegacyMixLayer::Two,
-        LegacyMixLayer::Three,
-    ];
-    let mut rng = OsRng;
-
-    // slap a random layer on it because legacy clients don't understand a concept of layerless mixnodes
-    // SAFETY: the slice is not empty so the unwrap is fine
-    #[allow(clippy::unwrap_used)]
-    let layer = layer_choices.choose(&mut rng).copied().unwrap();
-
-    LegacyMixNodeDetailsWithLayer {
-        bond_information: LegacyMixNodeBondWithLayer {
-            bond: MixNodeBond {
-                mix_id: nym_node.node_id(),
-                owner: nym_node.bond_information.owner.clone(),
-                original_pledge: nym_node.bond_information.original_pledge.clone(),
-                mix_node: MixNode {
-                    host: nym_node.bond_information.node.host.clone(),
-                    mix_port: description.mix_port(),
-                    verloc_port: description.verloc_port(),
-                    http_api_port: nym_node
-                        .bond_information
-                        .node
-                        .custom_http_port
-                        .unwrap_or(DEFAULT_NYM_NODE_HTTP_PORT),
-                    sphinx_key: description.host_information.keys.x25519.to_base58_string(),
-                    identity_key: nym_node.bond_information.node.identity_key.clone(),
-                    version: description.build_information.build_version.clone(),
-                },
-                proxy: None,
-                bonding_height: nym_node.bond_information.bonding_height,
-                is_unbonding: nym_node.bond_information.is_unbonding,
-            },
-            layer,
-        },
-        rewarding_details: nym_node.rewarding_details.clone(),
-        pending_changes: LegacyPendingMixNodeChanges {
-            pledge_change: nym_node.pending_changes.pledge_change,
-        },
-    }
-}
 
 // we want to mark the routes as deprecated in swagger, but still expose them
 #[allow(deprecated)]
@@ -232,25 +178,8 @@ async fn get_gateways(State(state): State<AppState>) -> Json<Vec<GatewayBond>> {
             continue;
         }
 
-        out.push(GatewayBond {
-            pledge_amount: nym_node.bond_information.original_pledge.clone(),
-            owner: nym_node.bond_information.owner.clone(),
-            block_height: nym_node.bond_information.bonding_height,
-            gateway: Gateway {
-                host: nym_node.bond_information.node.host.clone(),
-                mix_port: description.mix_port(),
-                clients_port: description.mixnet_websockets.ws_port,
-                location: description
-                    .auxiliary_details
-                    .location
-                    .map(|c| c.to_string())
-                    .unwrap_or_default(),
-                sphinx_key: description.host_information.keys.x25519.to_base58_string(),
-                identity_key: nym_node.bond_information.node.identity_key.clone(),
-                version: description.build_information.build_version.clone(),
-            },
-            proxy: None,
-        });
+        let node = to_legacy_gateway(nym_node, description);
+        out.push(node);
     }
 
     Json(out)
